:recycle: refactor creat folder endPoints
